refactor: split off utils
